Extract settings into separate window and auto-refresh on changes
- Move settings UI from a tab in LambdaPopup to its own SettingsWindow
- Revert LambdaPopup to original two-tab layout (Libraries/Search)
- Right-align Remove buttons with proper spacing in settings list
- Auto-refresh library data when repos are added/removed/toggled
- Support Enter key in repo URL input box
