+++ addin/lambda-boss/UI/SettingsWindow.xaml.cs
@@ -1,0 +1,150 @@
+using System.ComponentModel;
+using System.Windows;
+using System.Windows.Input;
+
+namespace LambdaBoss.UI;
+
+public partial class SettingsWindow
+{
+    public SettingsWindow()
+    {
+        InitializeComponent();
+        RefreshRepoList();
+        PreviewKeyDown += OnPreviewKeyDown;
+    }
+
+    /// <summary>
+    ///     Fired when the user changes settings (add/remove/toggle repos).
+    /// </summary>
+    public event EventHandler? SettingsChanged;
+
+    protected override void OnClosing(CancelEventArgs e)
+    {
+        e.Cancel = true;
+        Hide();
+    }
+
+    private void OnPreviewKeyDown(object sender, KeyEventArgs e)
+    {
+        if (e.Key == Key.Escape)
+        {
+            Hide();
+            e.Handled = true;
+        }
+    }
+
+    private void RefreshRepoList()
+    {
+        var settings = Settings.Current;
+        RepoList.ItemsSource = settings.Repos
+            .Select(r => new RepoDisplayItem
+            {
+                Url = r.Url,
+                Enabled = r.Enabled,
+                DisplayLabel = FormatRepoLabel(r),
+                LastFetchedLabel = r.LastFetched.HasValue
+                    ? $"Last fetched: {r.LastFetched.Value:yyyy-MM-dd HH:mm}"
+                    : "Never fetched"
+            })
+            .ToList();
+    }
+
+    private static string FormatRepoLabel(RepoConfig config)
+    {
+        try
+        {
+            var (owner, repo) = config.ParseOwnerRepo();
+            return $"{owner}/{repo}";
+        }
+        catch
+        {
+            return config.Url;
+        }
+    }
+
+    private void AddRepo()
+    {
+        var url = RepoUrlBox.Text.Trim();
+        if (string.IsNullOrEmpty(url))
+            return;
+
+        if (!url.StartsWith("https://github.com/", StringComparison.OrdinalIgnoreCase))
+        {
+            StatusText.Text = "URL must start with https://github.com/";
+            return;
+        }
+
+        var settings = Settings.Current;
+        if (!settings.AddRepo(url))
+        {
+            StatusText.Text = "Repository already exists";
+            return;
+        }
+
+        settings.Save();
+        RepoUrlBox.Text = "";
+        RefreshRepoList();
+        SettingsChanged?.Invoke(this, EventArgs.Empty);
+        StatusText.Text = $"Added {url}";
+    }
+
+    private void AddRepoButton_Click(object sender, RoutedEventArgs e)
+    {
+        AddRepo();
+    }
+
+    private void RepoUrlBox_KeyDown(object sender, KeyEventArgs e)
+    {
+        if (e.Key == Key.Enter)
+        {
+            AddRepo();
+            e.Handled = true;
+        }
+    }
+
+    private void RemoveRepoButton_Click(object sender, RoutedEventArgs e)
+    {
+        if (sender is not System.Windows.Controls.Button { Tag: string url })
+            return;
+
+        var settings = Settings.Current;
+        if (settings.Repos.Count <= 1)
+        {
+            StatusText.Text = "Cannot remove the last repository";
+            return;
+        }
+
+        if (settings.RemoveRepo(url))
+        {
+            settings.Save();
+            RefreshRepoList();
+            SettingsChanged?.Invoke(this, EventArgs.Empty);
+            StatusText.Text = $"Removed {url}";
+        }
+    }
+
+    private void RepoEnabledCheckbox_Click(object sender, RoutedEventArgs e)
+    {
+        if (sender is not System.Windows.Controls.CheckBox { DataContext: RepoDisplayItem item })
+            return;
+
+        var settings = Settings.Current;
+        var repo = settings.Repos.FirstOrDefault(r =>
+            string.Equals(r.Url.TrimEnd('/'), item.Url.TrimEnd('/'), StringComparison.OrdinalIgnoreCase));
+
+        if (repo != null)
+        {
+            repo.Enabled = item.Enabled;
+            settings.Save();
+            SettingsChanged?.Invoke(this, EventArgs.Empty);
+        }
+    }
+}
+
+internal class RepoDisplayItem
+{
+    public string Url { get; init; } = "";
+    public bool Enabled { get; set; }
+    public string DisplayLabel { get; init; } = "";
+    public string LastFetchedLabel { get; init; } = "";
+}
